+++ CHANGELOG.md
@@ -1,9 +1,24 @@
 ## Unreleased
 
 ### Added 
+* `InitiateKYCRecovery` operation ingestion
+* `CreateKYCRecoveryRequest` operation ingestion
+* `KYCRecovery` reviewable request ingestion
+* `GetKYCRecoveryRequests` to handle `/v3/kyc_recovery_requests`, `/v3/kyc_recovery_requests/{id}`
+* Docs on kyc recovery
 * `Asset` to withdraw request relationships
 * Filter by asset for `/v3/create_withdraw_requests` and `/v3/create_withdraw_requests/{id}`
 * Ability to include asset in responses`/v3/create_withdraw_requests` and `/v3/create_withdraw_requests/{id}`
+
+### Fixed
+
+* docs key value u32 value
+* add `trailing_digits_count` to asset creation request
+* Panic on `/v3/history` handling `ManageAccountSpecificRuleOp` details
+
+### Changed
+
+* `CurrentIngestVersion` increased
 
 ## 3.3.2